Backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
Flask-Migrate
Flask-Cors
python-dotenv
